Updated protocol.h to match enum names instead of uint32_t
git-svn-id: https://chan-sccp-b.svn.sourceforge.net/svnroot/chan-sccp-b/trunk@1297 fa19cda8-b258-4bce-aa1a-843e42da0b19

+++ src/sccp_labels.h
@@ -109,7 +109,7 @@
  * \brief Skinny LABEL Structure
  */
 static const struct skinny_label {
-        int label;
+        uint16_t label;
         const char * const text;
 } skinny_labels[] = {
 	{ SKINNY_LBL_EMPTY				, "Empty" },
